requirements.txt.txt

aiogram==3.7.0
qrcode[pil]==7.4.2
pillow==10.4.0